+++ appTools/ToolDrilling.py
@@ -311,9 +311,12 @@ class ToolDrilling(AppTool, Excellon):
         # populate Excellon preprocessor combobox list
         pp_list = []
         for name in self.app.preprocessors.keys():
-            # the HPGL preprocessor is only for Geometry not for Excellon job therefore don't add it
-            if name in ['hpgl', 'Paste_1', 'Check_points']:
+            # the HPGL preprocessor or Paste_ are not for Excellon job therefore don't add it
+            if name in ['hpgl', 'Paste_1']:
                 continue
+            elif name == 'Check_points':
+                pp_list = [name]
+                break
             pp_list.append(name)
         pp_list.sort()
         if 'default' in pp_list: